Move ProjectLink + AboutLink components to routes folder
