+++ src/components/index.js
@@ -3,8 +3,8 @@ export { App } from "./app/App.js";
 export { ProjectCardContainer } from "./projects/ProjectCardContainer";
 export { ProjectCard } from "./projects/ProjectCard";
 export { ProjectCardItem } from "./projects/ProjectCardItem";
-export { ProjectLink } from "./projects/ProjectLink";
+export { ProjectLink } from "./routes/ProjectLink";
 export { Footer } from "./footer/Footer";
-export { AboutLink } from "./about/AboutLink";
+export { AboutLink } from "./routes/AboutLink";
 export { About } from "./about/About";
 export { Routes } from "./routes/Routes";
